+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 - `serde::{Serialize, Deserialize}` impls (shouldn't break anything)
 - `MCaptcha` throws error when duration is 0
 - `Visitor` is changed to `AddVisitor`
+- `Master` packs a garbage collector to stop and get rid of inactive
+  `MCaptcha` actors
+- `Master` constructor accepts a parameter to configure GC(see previous
+  point) period